Refactor: Break Dashboard code to little reuseable components
